Define CompositeMenu
