+++ src/components/Menu/index.scss
@@ -1,16 +1,1 @@
-.menu {
-  @import "./TreeMenu/index";
-  @import "./ListMenu/index";
-  display: inline-block;
-  min-width: 16rem;
-  padding: 1rem;
-
-  & > *:not(:last-child) {
-    margin-bottom: .5rem;
-  }
-
-  & .button_group {
-    display: flex;
-    justify-content: center;
-  }
-}
+@import "CompositeMenu/index";
