+++ src/utils/component.js
@@ -1,5 +1,9 @@
+/* eslint react/no-multi-comp: 0 */
+
 import React from 'react';
+import ReactDOM from 'react-dom';
 import _ from 'lodash';
+import { modularizeRootNode } from './css';
 
 export const getRestProps = (componentInstance) => {
   const {
@@ -14,11 +18,29 @@ export const getRestProps = (componentInstance) => {
 export const mapProps = mapper => (
   Component => (
     class extends React.PureComponent {
-      static displayName = `mapProps(${Component.name})`;
+      static get name() {
+        return `mapProps(${Component.name})`;
+      }
 
       render() {
         return <Component {...mapper(this.props)} />;
       }
     }
   )
+);
+
+export const modularize = Component => (
+  class extends React.PureComponent {
+    static get name() {
+      return `modularize(${Component.name})`;
+    }
+
+    componentDidMount() {
+      modularizeRootNode(ReactDOM.findDOMNode(this).parentNode);
+    }
+
+    render() {
+      return <Component {...this.props} />;
+    }
+  }
 );
